+++ threedigrid/admin/crosssections/models.py
@@ -20,3 +20,20 @@ class CrossSections(Model):
     offset = ArrayField(type=float)
     count = ArrayField(type=float)
     tables = ArrayField(type=float)
+
+    def get_tabulated_cross_section_height_and_width(self, cross_section_pk):
+        """
+        Returns the heights and widths of a cross-section.
+
+        Args:
+            cross_section_pk: (int) the cross-section primary key
+
+        Returns:
+            (tuple([heights], [widths])) the heights and widths of the cross-section
+        """
+        offset = self.offset[cross_section_pk]
+        count = self.count[cross_section_pk]
+        return (
+            self.tables[0, offset : offset + count],
+            self.tables[1, offset : offset + count],
+        )